+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - Date and Time pickers popup are now positioned correctly in RTL mode ([#675](https://github.com/bdlukaa/fluent_ui/issues/675))
 - It's now possible to navigate through `AutoSuggestBox` items by long pressing arrow up and down keys
 - Do not clear focus scope after selecting an item in `AutoSuggestBox` ([#671](https://github.com/bdlukaa/fluent_ui/issues/671))
+- `AutoSuggestBox`'s `trailingIcon` now comes after the close button
+- **MINOR BREAK** `TextBox.clearGlobalKey` was remove, since it was not used
+- Add `AutoSuggestBox.unfocusedColor` and `TextFormBox.unfocusedColor`
 
 ## 4.1.3